fix(ci): 诊断改错误行过滤 + selftest 失败内联日志尾；修 %d/%lld 格式串 UB
- 诊断步骤先 grep 错误行（error:/fatal/undefined/LNK/MSB）再吐原始尾部，
  避免 GitHub 每 run 10 条注解上限把真正的错误行挤出可视区
- 三个 selftest 失败时把日志尾转 ::error:: 注解（匿名可读）
- Windows selftest 改 bash（Git Bash 语法统一）
- editor.h selftest 网格检查 qWarning 格式串 %d→%lld（qsizetype 是
  long long，旧写法是 varargs UB——GCC 11 已告警）

+++ editor.h
@@ -2708,9 +2708,9 @@ public:
             const QString art = Ascii::imageToText(simg, 8, 4);
             const QStringList lines = art.split(QLatin1Char('\n'));
             if (lines.size() != 4 || lines[0].size() != 8 || lines[3].size() != 8) {
-                qWarning("selftest FAIL: ascii art grid wrong (%d lines, sizes %d/%d)",
-                         int(lines.size()), lines.isEmpty() ? -1 : lines[0].size(),
-                         lines.size() < 4 ? -1 : lines[3].size());
+                qWarning("selftest FAIL: ascii art grid wrong (%lld lines, sizes %lld/%lld)",
+                         qint64(lines.size()), lines.isEmpty() ? -1 : qint64(lines[0].size()),
+                         lines.size() < 4 ? -1 : qint64(lines[3].size()));
                 return false;
             }
             if (lines[0].at(0) == QLatin1Char(' ') || lines[0].at(7) != QLatin1Char(' ')) {
